ref (back-views): séparation des vues pour plus de clarté
