+++ backend/api/views/__init__.py
@@ -1,0 +1,6 @@
+from .base_views import *
+from .documents_views import *
+from .equipement_views import *
+from .lieu_views import *
+from .intervention_views import *
+from .defaillance_views import *
